pyproject.toml
[tool.poetry]
name = "self-hosted"
version = "0.1.0"
description = ""
authors = ["Jane Jeon <git@janejeon.com>"]
readme = "README.md"
packages = [{include = "self_hosted"}]

[tool.poetry.dependencies]
python = "^3.11"
jinja2 = "^3.1.2"
requests = "^2.28.2"
python-dotenv = "^0.21.1"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"
pytest = "^7.2.1"
responses = "^0.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
